feat: support right-to-left languages (#110)
* feat: support right-to-left languages

This makes the Ray component styles sensitive to direction, such that components will render in the
expected orientation for right-to-left languages. Ideally, this would have been done using CSS' new
Logical Properties, but support for these is not pervasive yet, so style overrides are applied using
parent selectors wherever a left-right asymmetrical style is applied. The examples have been
extended to show every component with Hebrew right-to-left rendering.

Component modifiers that ended in "left" or "right" have been changed to "start"
and "end" respectively, to communicate the alignment relative to the expected flow of text in the
host language. All examples have been updated, but any clients using Ray will likely need to be
updated. Alternatively, we could temporarily alias make "left" == "start" and "right" == "end"
semantically, until we can fully deprecate those styles.

* revert: use padding-top instead of padding-block-start

Reverts the accidentally included use of logical properties to support language direction
sensitivity in the styles. These are not yet fully standardized or widely supported, and as this
change is not intended to support top-to-bottom languages, the change to the top padding was
irrelevant anyway.

* fix: reintroduce left and right rules as aliases of start and end

* fix: use css transform to flip breadcrumb arrow